Cambiar Float a Double en Volumen_de_la_esfera.hs

+++ src/Volumen_de_la_esfera.hs
@@ -6,15 +6,15 @@
 
 -- ---------------------------------------------------------------------
 -- Definir la función
---    volumenEsfera :: Float -> Float
+--    volumenEsfera :: Double -> Double
 -- tal que (volumenEsfera r) es el volumen de la esfera de radio r. Por
 -- ejemplo,
---    volumenEsfera 10  ==  4188.7905
+--    volumenEsfera 10  ==  4188.790204786391
 -- ---------------------------------------------------------------------
 
 {-# OPTIONS_GHC -fno-warn-type-defaults #-}
 
 module Volumen_de_la_esfera where
 
-volumenEsfera :: Float -> Float
+volumenEsfera :: Double -> Double
 volumenEsfera r = (4/3)*pi*r^3
